List all objects in an AWS bucket

aws_info() uses aws.s3::get_bucket_df() to return a tibble of stored objects in a remote bucket, like fs::dir_info() returns information on a local directory.

aws_ls() also uses aws.s3::get_bucket_df() but only returns the filenames within the bucket as a named fs_path character vector, like fs::dir_ls()) returns files in a local directory.

aws_info(bucket = aws_bucket(), prefix = NULL, ...)

aws_ls(bucket = aws_bucket(), prefix = NULL, ...)

aws_bucket(bucket = getOption("aws.bucket"), set = FALSE)

Arguments

bucket
Character string with the name of the bucket. If you use the same bucket frequently, you can set a default through an option named that can be retrieved with aws_bucket().

prefix
Character string that limits the response to keys that begin with the specified prefix.

...
Additional arguments passed to aws.s3::s3HTTP().

set
If TRUE, print instructions for setting the option.

Examples

aws_info("1000genomes", max = 10)

#> # A tibble: 10 x 4
#>    path                                       type      size time               
#>    <fs::path>                                 <fct> <fs::by> <dttm>             
#>  1 1000G_2504_high_coverage/1000G_2504_high_… file   947.83K 2019-04-22 00:28:31
#>  2 1000G_2504_high_coverage/additional_698_r… file   286.09K 2020-05-26 18:26:06
#>  3 1000G_2504_high_coverage/additional_698_r… file     95.2K 2020-05-26 18:26:51
#>  4 1000G_2504_high_coverage/additional_698_r… file     2.34K 2020-05-26 18:27:27
#>  5 1000G_2504_high_coverage/additional_698_r… file    14.12G 2020-05-19 17:55:54
#>  6 1000G_2504_high_coverage/additional_698_r… file    13.99G 2020-05-19 18:10:24
#>  7 1000G_2504_high_coverage/additional_698_r… file     15.2G 2020-05-19 17:33:12
#>  8 1000G_2504_high_coverage/additional_698_r… file    13.93G 2020-05-19 18:21:30
#>  9 1000G_2504_high_coverage/additional_698_r… file    14.56G 2020-05-19 18:27:45
#> 10 1000G_2504_high_coverage/additional_698_r… file    16.18G 2020-05-19 18:37:59

aws_ls("irs-form-990", max = 1)

#> 200931393493000150_public.xml