Create tar archive

A wrapper for an external tar command to create tar archives. Unlike utils::tar(), filepaths can be passed into the first argument with %>%.

tar_create(
  path,
  tar = NULL,
  compress = c("xz", "gzip", "bzip2", "none"),
  level = 6,
  extra = "",
  method = "tar"
)

Arguments

path
A character vector of files to archive.

tar
The path of the tarball file to create. If none is supplied, and only one path is, an archive with that basename() will be created in the same directory.

compress
Character string giving the type of compression to be used.

level
Integer compression level used with internal method, 1-9.

extra
An optional character vector passed to system() as arguments of the external tar command.

method
A character string specifying the external command to be used.

Value

The path of the created archive (invisibly).

Examples

.old_wd <- setwd(tempdir())
# Tar a file to same basename
tmp <- tempfile(fileext = ".csv")
write.csv(iris, tmp)
(txz <- tar_create(tmp))

#> /private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp6xYeU6/file69c7169e245.tar.xz

file.size(c(tmp, txz))

#> [1] 4821 1184

setwd(.old_wd)