List tar contents

tar_info() is the equivalent to the tar -tf command. It returns a data frame (like fs::dir_info()).

tar_ls() also uses tar -tf but only returns the filenames within the archive as a named fs_path character vector (like fs::dir_ls()).

tar_info(path, method = Sys.getenv("TAR"))

tar_ls(path, method = "internal")

tar_size(path, sum = TRUE, method = "tar")

Arguments

path
A character vector of tarball archives.

method
The method to be used. Passed to utils::tar(). An alternative is to use Sys.getenv("TAR").

sum
If TRUE (default) the summed size of archive contents is returned invisibly and a message comparing sizes is printed. If FALSE, the uncompressed size of all contents is returned as a numeric fs_bytes vector.

Examples

# List two files from zip
irs <- tempfile(fileext = ".csv")
write.csv(iris, irs)
mtc <- tempfile(fileext = ".csv")
write.csv(mtcars, mtc)
txz <- tempfile(fileext = ".tar")
tar_create(c(irs, mtc), txz)
tar_size(txz)

#> deflated: 0, compressed: 1.91K (InfY)

tar_info(txz)

#> # A tibble: 2 x 6
#>   path                          size permissions modification_time    user group
#>   <fs::path>                  <fs::> <fs::perms> <dttm>              <int> <chr>
#> 1 25 17:04 private var folde…     NA rw-rw-rw-   NA                      0 runn…
#> 2 25 17:04 private var folde…     NA rw-rw-rw-   NA                      0 runn…

tar_ls(txz)

#> private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp85G9e7/file65b1f330d30.csv
#> private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/Rtmp85G9e7/file65b50a6da50.csv