Extract zip contents

Decompress the contents of a zip archive and extract them to a single directory so they can be read. Compared to utils::unzip(), a directory must be provided in the dir directory; files are not automatically extracted to the current working directory. New file paths are returned invisibly.

tar_move(path, dir = NULL, files = NULL, junk = TRUE, method = "tar")

Arguments

path
A character vector of zip archives.

dir
The directory to extract files to (the equivalent of tar -C). It will be created if necessary with dir_create().

files
A character vector of paths to be extracted: the default is to extract all files. Use tar_ls() to list contents without extracting.

junk
If FALSE (default), preserve the directories of compressed files. If TRUE, use only the basename of the stored path when extracting; forward slashes in the tarball contents are counted and the maximum amount are passed to tar --strip-components.

method
The method to be used.

Value

The path to the extracted files (invisibly).

Examples

# Extract two files from tarball
tmp <- tempfile(fileext = ".csv")
write.csv(iris, tmp)
txz <- tar_create(tmp)
out <- tar_move(txz, dir = tempdir())
file.size(c(tmp, txz, out))

#> [1] 4821 1184   NA