Write the last value to disk

The value of the internal evaluation of a top-level R expression is always assigned to .Last.value before further processing (e.g., printing).

write_last(file = tempfile(), x = .Last.value, ...)

save_last(file = tempfile(), x = .Last.value, ...)

Arguments

file

File or connection to write to.

x

The object to write, usually left as base::.Last.value.

...

Additional arguments passed to the writing function (see Details).

Value

The created file path, invisibly.

Details

Four types of files are written, based on object class:

For data frames, a tab-separated file via readr::write_tsv().

For vectors, a newline-separated file via readr::write_lines().

For ggplots, a raster image (by default) via ggplot2::ggsave().

For other objects, an uncompressed data file via readr::write_rds().