List all objects in an AWS bucket

aws_info() uses aws.s3::get_bucket_df() to return a tibble of stored objects in a remote bucket, like fs::dir_info() returns information on a local directory.

aws_ls() also uses aws.s3::get_bucket_df() but only returns the filenames within the bucket as a named fs_path character vector, like fs::dir_ls()) returns files in a local directory.

aws_info(bucket = aws_bucket(), prefix = NULL, max = Inf, ...)

aws_ls(bucket = aws_bucket(), prefix = NULL, ...)

aws_bucket(bucket = getOption("aws.bucket"), set = FALSE)

Arguments

bucket

Character string with the name of the bucket. If you use the same bucket frequently, you can set a default through an option named that can be retrieved with aws_bucket().

prefix

Character string that limits the response to keys that begin with the specified prefix.

max

Number of objects to return.

...

Additional arguments passed to aws.s3::s3HTTP().

set

If TRUE, print instructions for setting the option.

Examples

aws_info("1000genomes", max = 10)
#> # A tibble: 10 × 4
#>    path                                        type     size time               
#>    <fs::path>                                  <fct> <fs::b> <dttm>             
#>  1 …ge/1000G_2504_high_coverage.sequence.index file  947.83K 2019-04-22 00:28:31
#>  2 …G_698_related_high_coverage.sequence.index file  286.09K 2020-05-26 18:26:06
#>  3 …130606_g1k_3202_samples_ped_population.txt file    95.2K 2020-05-26 18:26:51
#>  4 …2504plus698_high_cov_data_reuse_README.txt file    2.34K 2020-05-26 18:27:27
#>  5 …related/data/ERR3988761/HG00405.final.cram file   14.12G 2020-05-19 17:55:54
#>  6 …ed/data/ERR3988761/HG00405.final.cram.crai file    1.27M 2021-12-07 18:57:24
#>  7 …related/data/ERR3988762/HG00408.final.cram file   13.99G 2020-05-19 18:10:24
#>  8 …ed/data/ERR3988762/HG00408.final.cram.crai file    1.31M 2021-12-07 18:58:42
#>  9 …related/data/ERR3988763/HG00418.final.cram file    15.2G 2020-05-19 17:33:12
#> 10 …ed/data/ERR3988763/HG00418.final.cram.crai file    1.33M 2021-12-07 18:59:12
aws_ls("irs-form-990", max = 1)
#> character(0)